@@ -1,6 +1,6 @@
 # Generated by Django 4.2.20 on 2025-05-14 17:36
 
-import django_jsonform.models.fields
+import NEMO.fields
 from django.db import migrations
 
 
@@ -14,6 +14,6 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name="tool",
             name="_properties",
-            field=django_jsonform.models.fields.JSONField(blank=True, db_column="properties", null=True),
+            field=NEMO.fields.JsonField(blank=True, db_column="properties", null=True),
         ),
     ]
