@@ -50,7 +50,7 @@ class BillableItemMixin:
     CONSUMABLE = "consumable"
     MISSED_RESERVATION = "missed_reservation"
 
-    def get_customer(self) -> User:
+    def get_customer(self) -> Optional[User]:
         if self.get_real_type() in [
             BillableItemMixin.AREA_ACCESS,
             BillableItemMixin.REMOTE_WORK,
@@ -61,6 +61,7 @@ class BillableItemMixin:
             return self.user
         elif self.get_real_type() == BillableItemMixin.TRAINING:
             return self.trainee
+        return None
 
     def get_operator(self) -> Optional[User]:
         if self.get_real_type() == BillableItemMixin.AREA_ACCESS:
@@ -75,6 +76,7 @@ class BillableItemMixin:
             return self.merchant
         elif self.get_real_type() == BillableItemMixin.MISSED_RESERVATION:
             return self.creator
+        return None
 
     def get_item(self) -> str:
         if self.get_real_type() == BillableItemMixin.AREA_ACCESS:
@@ -90,12 +92,20 @@ class BillableItemMixin:
             return f"{self.consumable}{quantity}"
         elif self.get_real_type() == BillableItemMixin.MISSED_RESERVATION:
             return f"{self.tool or self.area} missed reservation"
+        return f"{self.__class__.__name__}"
 
     def waive(self, user: User):
-        if hasattr(self, "waived"):
+        if hasattr(self, "waived") and not self.waived:
             self.waived = True
             self.waived_by = user
             self.waived_on = timezone.now()
+            self.save(update_fields=["waived", "waived_by", "waived_on"])
+
+    def unwaive(self):
+        if hasattr(self, "waived") and self.waived:
+            self.waived = False
+            self.waived_by = None
+            self.waived_on = None
             self.save(update_fields=["waived", "waived_by", "waived_on"])
 
     def validate(self, user: User):
@@ -112,6 +122,7 @@ class BillableItemMixin:
             BillableItemMixin.MISSED_RESERVATION,
         ]:
             return self.start
+        return None
 
     def get_end(self) -> Optional[datetime.datetime]:
         if self.get_real_type() in [
@@ -123,18 +134,19 @@ class BillableItemMixin:
             return self.end
         elif self.get_real_type() in [BillableItemMixin.TRAINING, BillableItemMixin.CONSUMABLE]:
             return self.date
+        return None
 
     def can_be_adjusted(self, user: User):
         # determine if the given user can make an adjustment request for this charge
         from NEMO.views.customization import AdjustmentRequestsCustomization
         from NEMO.views.usage import get_managed_projects
 
-        pi_projects = get_managed_projects(user)
+        managed_projects = get_managed_projects(user)
 
         tool: Optional[Tool] = getattr(self, "tool", None)
         time_limit = AdjustmentRequestsCustomization.get_date_limit()
         time_limit_condition = not time_limit or time_limit <= self.get_end()
-        user_project_condition = self.get_customer() == user or self.project in pi_projects
+        user_project_condition = self.get_customer() == user or self.project in managed_projects
         operator_is_staff = self.get_operator() == user and user.is_staff_on_tool(tool)
         if self.get_real_type() == BillableItemMixin.AREA_ACCESS:
             access_enabled = AdjustmentRequestsCustomization.get_bool("adjustment_requests_area_access_enabled")
@@ -246,6 +258,7 @@ class BillableItemMixin:
             return "charged "
         elif self.get_real_type() == BillableItemMixin.MISSED_RESERVATION:
             return "created "
+        return ""
 
     def get_real_type(self) -> str:
         from NEMO.models import (
@@ -285,6 +298,7 @@ class BillableItemMixin:
             return self.get_real_type()
 
     def get_display(self, user: User = None) -> str:
+        project_display = f", charged to {self.project}"
         customer_display = f" for {self.get_customer()}"
         operator_display = f" {self.get_operator_action()}by {self.get_operator()}"
         user_display = ""
@@ -298,7 +312,7 @@ class BillableItemMixin:
             charge_time = f" from {format_datetime(self.get_start(), DTF)} to {format_datetime(self.get_end(), DTF)}"
         elif self.get_start() or self.get_end():
             charge_time = f" on {format_datetime(self.get_start() or self.get_end(), DTF)}"
-        return f"{self.get_item()}{user_display}{charge_time}"
+        return f"{self.get_item()}{user_display}{charge_time}{project_display}"
 
 
 class RecurrenceMixin:
