@@ -39,6 +39,7 @@ from NEMO.models import (
     Interlock,
     InterlockCard,
     InterlockCardCategory,
+    MembershipHistory,
     PhysicalAccessLevel,
     Project,
     ProjectDiscipline,
@@ -46,6 +47,7 @@ from NEMO.models import (
     Qualification,
     RecurringConsumableCharge,
     Reservation,
+    ReservationQuestions,
     Resource,
     ScheduledOutage,
     StaffAssistanceRequest,
@@ -55,9 +57,12 @@ from NEMO.models import (
     TemporaryPhysicalAccessRequest,
     Tool,
     ToolCredentials,
+    ToolUsageCounter,
+    ToolUsageQuestions,
     TrainingSession,
     UsageEvent,
     User,
+    UserCalendarToolList,
     UserDocuments,
     UserPreferences,
 )
@@ -166,6 +171,8 @@ class UserSerializer(FlexFieldsSerializerMixin, ModelSerializer):
         expandable_fields = {
             "projects": ("NEMO.serializers.ProjectSerializer", {"many": True}),
             "managed_projects": ("NEMO.serializers.ProjectSerializer", {"many": True}),
+            "managed_accounts": ("NEMO.serializers.AccountSerializer", {"many": True}),
+            "managed_users": ("NEMO.serializers.UserSerializer", {"many": True}),
             "groups": ("NEMO.serializers.GroupSerializer", {"many": True}),
             "user_documents": ("NEMO.serializers.UserDocumentSerializer", {"many": True}),
             "user_permissions": ("NEMO.serializers.PermissionSerializer", {"many": True}),
@@ -194,6 +201,16 @@ class UserDocumentSerializer(FlexFieldsSerializerMixin, ModelSerializer):
         fields = "__all__"
         expandable_fields = {
             "user": "NEMO.serializers.UserSerializer",
+        }
+
+
+class UserCalendarToolListSerializer(FlexFieldsSerializerMixin, ModelSerializer):
+    class Meta:
+        model = UserCalendarToolList
+        fields = "__all__"
+        expandable_fields = {
+            "user": "NEMO.serializers.UserSerializer",
+            "tools": ("NEMO.serializers.ToolSerializer", {"many": True}),
         }
 
 
@@ -270,7 +287,6 @@ class ToolSerializer(FlexFieldsSerializerMixin, ModelSerializer):
             "_superusers": ("NEMO.serializers.UserSerializer", {"many": True}),
             "_requires_area_access": "NEMO.serializers.AreaSerializer",
             "project": "NEMO.serializers.ProjectSerializer",
-            "descendant": "NEMO.serializers.ReservationSerializer",
         }
 
 
@@ -318,6 +334,17 @@ class ReservationSerializer(FlexFieldsSerializerMixin, ModelSerializer):
             "configuration_options": ("NEMO.serializers.ConfigurationOptionSerializer", {"many": True}),
             "validated_by": "NEMO.serializers.UserSerializer",
             "waived_by": "NEMO.serializers.UserSerializer",
+        }
+
+
+class ReservationQuestionsSerializer(FlexFieldsSerializerMixin, ModelSerializer):
+    class Meta:
+        model = ReservationQuestions
+        fields = "__all__"
+        expandable_fields = {
+            "only_for_tools": ("NEMO.serializers.ToolSerializer", {"many": True}),
+            "only_for_areas": ("NEMO.serializers.AreaSerializer", {"many": True}),
+            "only_for_projects": ("NEMO.serializers.ProjectSerializer", {"many": True}),
         }
 
 
@@ -536,6 +563,31 @@ class AdjustmentRequestSerializer(FlexFieldsSerializerMixin, ModelSerializer):
             "reviewer": "NEMO.serializers.UserSerializer",
             "item_type": "NEMO.serializers.ContentTypeSerializer",
             "applied_by": "NEMO.serializers.UserSerializer",
+            "new_project": "NEMO.serializers.ProjectSerializer",
+            "item_tool": "NEMO.serializers.ToolSerializer",
+            "item_area": "NEMO.serializers.AreaSerializer",
+        }
+
+
+class ToolUsageQuestionsSerializer(FlexFieldsSerializerMixin, ModelSerializer):
+    class Meta:
+        model = ToolUsageQuestions
+        fields = "__all__"
+        expandable_fields = {
+            "only_for_tools": ("NEMO.serializers.ToolSerializer", {"many": True}),
+            "only_for_projects": ("NEMO.serializers.ProjectSerializer", {"many": True}),
+            "only_for_users": ("NEMO.serializers.UserSerializer", {"many": True}),
+            "only_for_groups": ("NEMO.serializers.GroupSerializer", {"many": True}),
+        }
+
+
+class ToolUsageCounterSerializer(FlexFieldsSerializerMixin, ModelSerializer):
+    class Meta:
+        model = ToolUsageCounter
+        fields = "__all__"
+        expandable_fields = {
+            "tool": "NEMO.serializers.ToolSerializer",
+            "last_reset_by": "NEMO.serializers.UserSerializer",
         }
 
 
@@ -567,6 +619,17 @@ class StaffAssistanceRequestSerializer(FlexFieldsSerializerMixin, ModelSerialize
         expandable_fields = {
             "tool": "NEMO.serializers.ToolSerializer",
             "user": "NEMO.serializers.UserSerializer",
+        }
+
+
+class MembershipHistorySerializer(FlexFieldsSerializerMixin, ModelSerializer):
+    class Meta:
+        model = MembershipHistory
+        fields = "__all__"
+        expandable_fields = {
+            "parent_content_type": "NEMO.serializers.ContentTypeSerializer",
+            "child_content_type": "NEMO.serializers.ContentTypeSerializer",
+            "authorizer": "NEMO.serializers.UserSerializer",
         }
 
 
