@@ -16,10 +16,10 @@ from NEMO.models import (
     ScheduledOutage,
     User,
 )
-from NEMO.tests.test_utilities import login_as, login_as_user
+from NEMO.tests.test_utilities import NEMOTestCaseMixin
 
 
-class AreaReservationTestCase(TransactionTestCase):
+class AreaReservationTestCase(NEMOTestCaseMixin, TransactionTestCase):
     area: Area = None
     area_access_level = None
     owner: User = None
@@ -62,7 +62,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, user)
+        self.login_as(user)
 
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
@@ -82,7 +82,7 @@ class AreaReservationTestCase(TransactionTestCase):
         user.access_expiration = None
         user.training_required = False
         user.save()
-        login_as(self.client, user)
+        self.login_as(user)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertNotContains(
@@ -99,7 +99,7 @@ class AreaReservationTestCase(TransactionTestCase):
         )
 
         user.physical_access_levels.add(area_access_level)
-        login_as(self.client, user)
+        self.login_as(user)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertNotContains(
@@ -117,7 +117,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = base_start - timedelta(hours=1)
         data = self.get_reservation_data(base_start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
 
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 400)
@@ -172,7 +172,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertContains(response, "You may not create reservations further than 2 days from now for this area.")
@@ -186,7 +186,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertContains(
@@ -301,7 +301,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(Reservation.objects.all().count(), 1)
@@ -312,7 +312,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
 
         second_project = Project.objects.create(name="project2", account=Account.objects.get(name="account1"))
         consumer.projects.add(second_project)
@@ -345,7 +345,7 @@ class AreaReservationTestCase(TransactionTestCase):
         data = self.get_reservation_data(start, end, area)
         data["impersonate"] = consumer.id
 
-        login_as(self.client, staff)
+        self.login_as(staff)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         self.assertEqual(Reservation.objects.all().count(), 1)
@@ -359,7 +359,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         reservation = Reservation.objects.get(area=area)
@@ -484,7 +484,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         reservation = Reservation.objects.get(area=area)
@@ -566,7 +566,7 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         reservation = Reservation.objects.get(area=area)
@@ -581,7 +581,7 @@ class AreaReservationTestCase(TransactionTestCase):
         self.assertEqual(response.status_code, 404)
 
         # test non staff user trying to cancel reservation
-        login_as_user(self.client)
+        self.login_as_user()
         response = self.client.post(
             reverse("cancel_reservation", kwargs={"reservation_id": reservation.id}), {}, follow=True
         )
@@ -596,7 +596,7 @@ class AreaReservationTestCase(TransactionTestCase):
         self.assertEqual(response.status_code, 400)
         self.assertEqual("You may not cancel reservations that you do not own.", response.content.decode())
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
 
         # test cancel missed reservation
         missed_resa = Reservation.objects.create(
@@ -670,7 +670,7 @@ class AreaReservationTestCase(TransactionTestCase):
             creator=consumer,
             short_notice=False,
         )
-        login_as(self.client, staff)
+        self.login_as(staff)
         response = self.client.post(
             reverse("cancel_reservation", kwargs={"reservation_id": other_resa.id}), {}, follow=True
         )
@@ -691,14 +691,14 @@ class AreaReservationTestCase(TransactionTestCase):
         end = start + timedelta(hours=1)
         data = self.get_reservation_data(start, end, area)
 
-        login_as(self.client, consumer)
+        self.login_as(consumer)
         response = self.client.post(reverse("create_reservation"), data, follow=True)
         self.assertEqual(response.status_code, 200)
         reservation = Reservation.objects.get(area=area)
         self.assertTrue(reservation.id)
 
         # anybody that is logged in can see reservation details
-        login_as_user(self.client)
+        self.login_as_user()
 
         # post should fail
         response = self.client.post(reverse("reservation_details", kwargs={"reservation_id": 999}), {}, follow=True)
@@ -712,6 +712,60 @@ class AreaReservationTestCase(TransactionTestCase):
             reverse("reservation_details", kwargs={"reservation_id": reservation.id}), {}, follow=True
         )
         self.assertEqual(response.status_code, 200)
+
+    def test_change_reservation_note(self):
+        # create reservation
+        start = datetime.now() + timedelta(hours=1)
+        end = start + timedelta(hours=1)
+        data = self.get_reservation_data(start, end, area)
+
+        self.login_as(consumer)
+        response = self.client.post(reverse("create_reservation"), data, follow=True)
+        self.assertEqual(response.status_code, 200)
+        reservation = Reservation.objects.get(area=area)
+        self.assertTrue(reservation.id)
+
+        # Consumer or staff should see the note field
+        response = self.client.get(
+            reverse("reservation_details", kwargs={"reservation_id": reservation.id}), {}, follow=True
+        )
+        self.assertContains(response, "Note")
+
+        self.login_as_staff()
+        response = self.client.get(
+            reverse("reservation_details", kwargs={"reservation_id": reservation.id}), {}, follow=True
+        )
+        self.assertContains(response, "Note")
+
+        # non-staff and non reservation owner cannot see the reservation note
+        self.login_as_user()
+        response = self.client.get(
+            reverse("reservation_details", kwargs={"reservation_id": reservation.id}), {}, follow=True
+        )
+        self.assertNotContains(response, "Note")
+        # they also cannot change the reservation note
+        response = self.client.post(
+            reverse("change_reservation_note", kwargs={"reservation_id": reservation.id}),
+            {"note": "this is a note"},
+            follow=True,
+        )
+        self.assertEqual(response.status_code, 400)
+
+        # owner can change, but not with wrong id
+        self.login_as(consumer)
+        # test wrong id
+        response = self.client.post(reverse("change_reservation_note", kwargs={"reservation_id": 999}), {}, follow=True)
+        self.assertEqual(response.status_code, 404)
+        # try again with correct id
+        response = self.client.post(
+            reverse("change_reservation_note", kwargs={"reservation_id": reservation.id}),
+            {"note": "this is a note"},
+            follow=True,
+        )
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(Reservation.objects.get(pk=reservation.id).note, "this is a note")
+
+        reservation.delete()
 
     def test_reservation_with_area_configuration(self):
         # TODO: create those tests
