@@ -1,5 +1,7 @@
 from datetime import date, datetime, time, timedelta
 
+from NEMO.tests.test_utilities import NEMOTestCaseMixin
+
 try:
     import zoneinfo
 except ImportError:
@@ -14,7 +16,7 @@ from django.utils.timezone import make_aware
 from NEMO.utilities import format_daterange, format_datetime, get_duration_with_off_schedule
 
 
-class FormatTestCase(TestCase):
+class FormatTestCase(NEMOTestCaseMixin, TestCase):
     def test_format_daterange(self):
         self.assertTrue(format_datetime())
         today = datetime.today()
@@ -55,8 +57,8 @@ class FormatTestCase(TestCase):
         end_date = date(2022, 2, 11)
         self.assertEqual(format_daterange(start_date, end_date, d_format=d_format), f"from 02/11/2022 to 02/11/2022")
 
-        tz = zoneinfo.ZoneInfo("US/Pacific")
-        self.assertEqual(settings.TIME_ZONE, "US/Eastern")
+        tz = zoneinfo.ZoneInfo("America/Los_Angeles")
+        self.assertEqual(settings.TIME_ZONE, "America/New_York")
         start_tz = make_aware(datetime(2022, 2, 11, 5, 0, 0), tz)  # 5AM Pacific => 8AM Eastern
         end_tz = start_tz + timedelta(days=2)
         self.assertNotEqual(
