@@ -15,6 +15,7 @@ from pymodbus.pdu.bit_message import (
 # This method will be used by the mock to replace socket.send
 # In the stanford interlock case, it will return a list of byte
 from NEMO.models import Interlock, InterlockCard, InterlockCardCategory, Tool, User
+from NEMO.tests.test_utilities import NEMOTestCaseMixin
 
 server1 = "server1.nist.gov"
 server2 = "https://server2.nist.gov"
@@ -42,7 +43,7 @@ def mocked_modbus_client(*args, **kwargs):
     return MockModbusClient(*args, **kwargs)
 
 
-class ModbusInterlockTestCase(TestCase):
+class ModbusInterlockTestCase(NEMOTestCaseMixin, TestCase):
     tool: Tool = None
     wrong_response_interlock: Interlock = None
     bad_interlock: Interlock = None
