@@ -6,13 +6,14 @@ from django.test import TestCase
 from django.utils import timezone
 
 from NEMO.models import Account, EmailLog, Project, Qualification, Tool, UsageEvent, User
+from NEMO.tests.test_utilities import NEMOTestCaseMixin
 from NEMO.views.customization import EmailsCustomization, ToolCustomization
 from NEMO.views.timed_services import do_manage_tool_qualifications
 
 
 @patch("django.core.files.storage.FileSystemStorage.exists")
 @patch("django.core.files.storage.FileSystemStorage._open")
-class ToolQualificationTestCase(TestCase):
+class ToolQualificationTestCase(NEMOTestCaseMixin, TestCase):
     def setUp(self):
         self.manager: User = User.objects.create(
             username="manager",
