@@ -5,7 +5,8 @@ import requests
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
-from django.shortcuts import render
+from django.shortcuts import redirect, render
+from django.template import Context, Template
 from django.urls import Resolver404, resolve
 from django.utils import timezone
 from django.views.decorators.http import require_GET
@@ -13,7 +14,7 @@ from django.views.decorators.http import require_GET
 from NEMO.models import Alert, LandingPageChoice, Reservation, Resource, UsageEvent, User
 from NEMO.views.alerts import mark_alerts_as_expired
 from NEMO.views.area_access import able_to_self_log_in_to_area, able_to_self_log_out_of_area
-from NEMO.views.customization import UserCustomization
+from NEMO.views.customization import ApplicationCustomization, UserCustomization
 from NEMO.views.notifications import delete_expired_notifications
 
 
@@ -23,6 +24,8 @@ def landing(request):
     user: User = request.user
     mark_alerts_as_expired()
     delete_expired_notifications()
+    if not request.META.get("HTTP_REFERER") and user.get_preferences().login_redirect_url:
+        return redirect(user.get_preferences().login_redirect_url)
     usage_events = UsageEvent.objects.filter(operator=user.id, end=None).prefetch_related("tool", "project")
     tools_in_use = [u.tool.tool_or_parent_id() for u in usage_events]
     fifteen_minutes_from_now = timezone.now() + timedelta(minutes=15)
@@ -78,6 +81,9 @@ def landing(request):
         "self_log_in": able_to_self_log_in_to_area(request.user),
         "self_log_out": able_to_self_log_out_of_area(request.user),
         "script_name": settings.FORCE_SCRIPT_NAME,
+        "facility_rules_required_message": Template(
+            ApplicationCustomization.get("facility_rules_required_message")
+        ).render(Context()),
     }
     return render(request, "landing.html", dictionary)
 
