@@ -1,5 +1,6 @@
 from datetime import datetime, timedelta
 from itertools import chain
+from typing import Any, Dict
 
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
@@ -13,11 +14,7 @@ from NEMO.exceptions import ProjectChargeException, RequiredUnansweredQuestionsE
 from NEMO.models import Area, Project, Reservation, ReservationItemType, ScheduledOutage, Tool, User
 from NEMO.policy import policy_class as policy
 from NEMO.utilities import beginning_of_the_day, end_of_the_day, localize
-from NEMO.views.calendar import (
-    extract_reservation_questions,
-    render_reservation_questions,
-    set_reservation_configuration,
-)
+from NEMO.views.calendar import set_reservation_configuration
 from NEMO.views.customization import CalendarCustomization
 
 
@@ -26,7 +23,7 @@ from NEMO.views.customization import CalendarCustomization
 def choose_item(request, next_page):
     user: User = request.user
     tools = Tool.objects.filter(visible=True).order_by("_category", "name")
-    dictionary = {
+    dictionary: Dict[str, Any] = {
         "tools": list(tools),
         "areas": [],
     }
@@ -69,7 +66,7 @@ def new_reservation(request, item_type, item_id, date=None):
     dictionary = {
         "item": item,
         "item_type": item_type.value,
-        "date": date,
+        "date": parse_date(date) if date else None,
         "item_reservation_times": list(
             Reservation.objects.filter(**{item_type.value: item}).filter(
                 cancelled=False, missed=False, shortened=False, start__gte=timezone.now()
@@ -130,9 +127,9 @@ def make_reservation(request):
         if not user.is_staff and not user.is_staff_on_tool(item):
             return render(request, "mobile/error.html", {"message": "You must specify a project for your reservation"})
 
-    reservation_questions = render_reservation_questions(item_type, item.id, reservation.project)
+    dynamic_forms = item.get_reservation_questions(reservation.project)
     tool_config = item_type == ReservationItemType.TOOL and item.is_configurable()
-    needs_extra_config = reservation_questions or tool_config
+    needs_extra_config = dynamic_forms or tool_config
     if needs_extra_config and not request.POST.get("configured") == "true":
         dictionary = {
             "request_date": request.POST["date"],
@@ -142,7 +139,7 @@ def make_reservation(request):
         }
         if item_type == ReservationItemType.TOOL:
             dictionary.update(item.get_configuration_information(user=user, start=reservation.start))
-        dictionary["reservation_questions"] = reservation_questions
+        dictionary["reservation_questions"] = dynamic_forms.render()
         return render(request, "mobile/reservation_extra.html", dictionary)
 
     set_reservation_configuration(reservation, request)
@@ -152,7 +149,7 @@ def make_reservation(request):
 
     # Reservation questions if applicable
     try:
-        reservation.question_data = extract_reservation_questions(request, item_type, item.id, reservation.project)
+        reservation.question_data = dynamic_forms.extract(request)
     except RequiredUnansweredQuestionsException as e:
         return render(request, "mobile/error.html", {"message": str(e)})
 
