@@ -5,6 +5,7 @@ from typing import Callable, List, Set
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.db.models import F, Q
+from django.http import HttpResponseBadRequest
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 from django.views.decorators.http import require_GET
@@ -33,6 +34,7 @@ from NEMO.utilities import (
     get_day_timeframe,
     get_month_timeframe,
     month_list,
+    quiet_int,
 )
 from NEMO.views.api_billing import (
     BillableItem,
@@ -43,7 +45,7 @@ from NEMO.views.api_billing import (
     billable_items_training_sessions,
     billable_items_usage_events,
 )
-from NEMO.views.customization import AdjustmentRequestsCustomization, ProjectsAccountsCustomization
+from NEMO.views.customization import AdjustmentRequestsCustomization, ProjectsAccountsCustomization, ToolCustomization
 
 logger = getLogger(__name__)
 
@@ -106,12 +108,16 @@ def user_usage(request):
     trainee_filter = Q(trainee=user) | Q(project__in=user_managed_projects)
     show_only_my_usage = user_managed_projects and request.GET.get("show_only_my_usage", "enabled") == "enabled"
     csv_export = bool(request.GET.get("csv", False))
-    if show_only_my_usage:
-        # Forcing to be user only
-        customer_filter &= Q(customer=user)
-        user_filter &= Q(user=user)
-        trainee_filter &= Q(trainee=user)
-        csv_export = bool(request.GET.get("csv", False))
+    managed_user_id = quiet_int(request.GET.get("managed_user"))
+    if managed_user_id and not managed_user_id in user.managed_users.values_list("id", flat=True):
+        return HttpResponseBadRequest("You are not allowed to see this user's usage.")
+    if show_only_my_usage or managed_user_id:
+        # Forcing to be user only (either current user or the one selected)
+        forced_user_id = managed_user_id or user.id
+        customer_filter = Q(customer_id=forced_user_id)
+        user_filter = Q(user_id=forced_user_id)
+        trainee_filter = Q(trainee_id=forced_user_id)
+        show_only_my_usage = not managed_user_id
     return usage(
         request,
         usage_filter=user_filter,
@@ -129,14 +135,17 @@ def user_usage(request):
 @any_staff_required
 @require_GET
 def staff_usage(request):
-    user: User = request.user
-    usage_filter = Q(operator=user) & ~Q(user=F("operator"))
-    area_access_filter = Q(staff_charge__staff_member=user)
-    staff_charges_filter = Q(staff_member=user)
-    consumable_filter = Q(merchant=user)
-    user_filter = Q(pk__in=[])
-    trainee_filter = Q(trainer=user)
     csv_export = bool(request.GET.get("csv", False))
+    managed_user_id = quiet_int(request.GET.get("managed_user"))
+    if managed_user_id and not managed_user_id in request.user.managed_users.values_list("id", flat=True):
+        return HttpResponseBadRequest("You are not allowed to see this user's usage.")
+    user_id = managed_user_id or request.user.id
+    usage_filter = Q(operator_id=user_id) & ~Q(user=F("operator"))
+    area_access_filter = Q(staff_charge__staff_member_id=user_id)
+    staff_charges_filter = Q(staff_member_id=user_id)
+    consumable_filter = Q(merchant_id=user_id)
+    user_filter = Q(pk__in=[])
+    trainee_filter = Q(trainer_id=user_id)
     return usage(
         request,
         usage_filter=usage_filter,
@@ -165,6 +174,14 @@ def usage(
 ):
     user: User = request.user
     base_dictionary, start_date, end_date, kind, identifier = date_parameters_dictionary(request, get_month_timeframe)
+    # Preloading user's managed projects' and managed users`
+    base_dictionary["user"] = (
+        User.objects.filter(id=user.id).prefetch_related("managed_projects", "managed_users").first()
+    )
+    selected_managed_user_id = request.GET.get("managed_user")
+    base_dictionary["selected_managed_user"] = selected_managed_user_id
+    if selected_managed_user_id:
+        base_dictionary["explicitly_display_customer"] = True
     project_id = request.GET.get("project") or request.GET.get("pi_project")
     if user_managed_projects:
         base_dictionary["selected_project"] = "all"
@@ -215,10 +232,10 @@ def usage(
                     + list(staff_charges.values_list("project", flat=True))
                     + list(training_sessions.values_list("project", flat=True))
                 )
-            ),
+            ).order_by("account__name"),
         }
         if user_managed_projects:
-            dictionary["pi_projects"] = user_managed_projects
+            dictionary["managed_projects"] = sorted(user_managed_projects, key=lambda x: x.account.name)
             dictionary["show_only_my_usage"] = show_only_my_usage
         dictionary["no_charges"] = not (
             dictionary["area_access"]
@@ -238,8 +255,10 @@ def billing(request):
     base_dictionary, start_date, end_date, kind, identifier = date_parameters_dictionary(request, get_month_timeframe)
     if not base_dictionary["billing_service"]:
         return redirect("user_usage")
-    user_project_applications = list(user.active_projects().values_list("application_identifier", flat=True)) + list(
-        user.managed_projects.values_list("application_identifier", flat=True)
+    user_project_applications = (
+        list(user.active_projects().values_list("application_identifier", flat=True))
+        + list(user.managed_projects.values_list("application_identifier", flat=True))
+        + list(user.managed_accounts.values_list("project__application_identifier", flat=True))
     )
     formatted_applications = ",".join(map(str, set(user_project_applications)))
     try:
@@ -254,6 +273,8 @@ def billing(request):
 @require_GET
 def project_usage(request):
     base_dictionary, start_date, end_date, kind, identifier = date_parameters_dictionary(request, get_day_timeframe)
+    # Preloading user's managed projects'
+    base_dictionary["user"] = User.objects.filter(id=request.user.id).prefetch_related("managed_projects").first()
 
     area_access, consumables, missed_reservations, staff_charges, training_sessions, usage_events = (
         None,
@@ -273,7 +294,7 @@ def project_usage(request):
     selected_project_type = request.GET.get("project_type")
 
     try:
-        if kind == "application":
+        if kind == "projectapplication":
             projects = Project.objects.filter(application_identifier=identifier)
             selection = identifier
         elif kind == "project":
@@ -394,7 +415,7 @@ def project_billing(request):
     formatted_applications = None
     selection = ""
     try:
-        if kind == "application":
+        if kind == "projectapplication":
             formatted_applications = identifier
             selection = identifier
         elif kind == "project":
@@ -452,7 +473,7 @@ def is_user_pi(user: User, latest_pis_data, activity, user_managed_applications:
 
 def billing_dict(start_date, end_date, user, formatted_applications, project_id=None, account_id=None, force_pi=False):
     # The parameter force_pi allows us to display information as if the user was the project pi
-    # This is useful on the admin project billing page tp display other project users for example
+    # This is useful on the admin project billing page to display other project users for example
     dictionary = {}
 
     billing_service = get_billing_service()
@@ -484,7 +505,15 @@ def billing_dict(start_date, end_date, user, formatted_applications, project_id=
     # Construct a tree of account, application, project, and member total spending
     cost_activities_tree = {}
     user_managed_applications = (
-        [project.application_identifier for project in user.managed_projects.all()] if not force_pi else []
+        [
+            project.application_identifier
+            for project in Project.objects.filter(
+                Q(id__in=user.managed_projects.values_list("id", flat=True))
+                | Q(account_id__in=user.managed_accounts.values_list("id", flat=True))
+            )
+        ]
+        if not force_pi
+        else []
     )
     for activity in cost_activity_data:
         if (project_id and activity["project_id"] != str(project_id)) or (
@@ -537,6 +566,7 @@ def csv_export_response(
     table_result.add_header(("user", "User"))
     table_result.add_header(("name", "Item"))
     table_result.add_header(("details", "Details"))
+    table_result.add_header(("account", "Account"))
     table_result.add_header(("project", "Project"))
     if user.is_any_part_of_staff:
         table_result.add_header(
@@ -545,6 +575,7 @@ def csv_export_response(
     table_result.add_header(("start", "Start time"))
     table_result.add_header(("end", "End time"))
     table_result.add_header(("quantity", "Quantity"))
+    table_result.add_header(("note", "Note"))
     data: List[BillableItem] = []
     data.extend(billable_items_missed_reservations(missed_reservations))
     data.extend(billable_items_consumable_withdrawals(consumables))
@@ -553,7 +584,16 @@ def csv_export_response(
     data.extend(billable_items_area_access_records(area_access))
     data.extend(billable_items_usage_events(usage_events))
     for billable_item in data:
-        table_result.add_row(vars(billable_item))
+        row = vars(billable_item)
+        if billable_item.type == "staff_charge" and billable_item.item:
+            row["note"] = billable_item.item.note
+        if (
+            ToolCustomization.get_bool("tool_control_note_show")
+            and billable_item.type == "tool_usage"
+            and billable_item.item
+        ):
+            row["note"] = billable_item.item.note
+        table_result.add_row(row)
     response = table_result.to_csv()
     filename = f"usage_export_{export_format_datetime()}.csv"
     response["Content-Disposition"] = f'attachment; filename="{filename}"'
@@ -562,7 +602,14 @@ def csv_export_response(
 
 def get_managed_projects(user: User) -> Set[Project]:
     # This function will get managed projects from NEMO and also attempt to get them from billing service
-    managed_projects = set(list(user.managed_projects.all()))
+    managed_projects = set(
+        list(
+            Project.objects.filter(
+                Q(id__in=user.managed_projects.values_list("id", flat=True))
+                | Q(account_id__in=user.managed_accounts.values_list("id", flat=True))
+            )
+        )
+    )
     billing_service = get_billing_service()
     if billing_service.get("available", False):
         # if we have a billing service, use it to determine project lead
